refactor: client dto 구조 변경 kakao, apple로 구분
